Removed the service, now all the notification stuff is handeled in the new class called NotificationBroadcast. There is a static function there to allow a caller to set up an intent properly.
Added a class called MidnightBroadcast, it is set up to be called at
midnight to notify everyone to update there screens, it will also call
the code to fix any midnight problems.

Fixed an email bug, we will see if that is the end of it.

Fixed all service bugs by removing the service

Change-Id: I73b5170529cb6e51586b5820c8c1b71c73391fc7